del: 사용하지 않는 자바 파일 삭제

+++ src/main/java/com/mgnt/ticketing/controller/user/dto/response/GetBalanceResponse.java
@@ -1,6 +1,11 @@
 package com.mgnt.ticketing.controller.user.dto.response;
 
+import java.math.BigDecimal;
+
+/**
+ * 사용자 잔액 조회 응답 DTO
+ */
 public record GetBalanceResponse(
-        int balance
+        BigDecimal balance
 ) {
 }
